renames ui/station to ui/stations
