+++ lib/ui/screens/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:lingo_jam/ui/stations/stations_view.dart';
 import 'package:lingo_jam/ui/style/style_constants.dart';
 import 'package:lingo_jam/ui/widgets/player.dart';
-import 'package:lingo_jam/ui/station/stations_view.dart';
 import 'package:lingo_jam/ui/widgets/bottom_bar.dart';
 
 class HomeScreen extends StatelessWidget {
