+++ lib/ui/stations/stations_collections.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:lingo_jam/ui/stations/stations_collection.dart';
 import 'package:provider/provider.dart';
 import 'package:lingo_jam/core/providers/stations_state.dart';
-import 'package:lingo_jam/ui/station/stations_collection.dart';
 import 'package:lingo_jam/services/stations_collection_service.dart'
     as services_sc;
